backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
fastapi-swagger
PyJWT